@@ -3,9 +3,9 @@
 /// Foundation for Core Entry Store
 library core_persistence_local;
 
-export 'src/core_persistence_adapter_local.dart';
-export 'src/core_persistence_adapter_local_factory.dart';
-export 'src/core_persistence_provider_local.dart';
-export 'src/persistence_provider_local/persistence_adapter_local.dart';
-export 'src/persistence_provider_local/persistence_adapter_local_factory.dart';
-export 'src/persistence_provider_local/persistence_provider_local.dart';
+export 'src/core_persistence_local_adapter.dart';
+export 'src/core_persistence_local_adapter_factory.dart';
+export 'src/core_persistence_local_provider.dart';
+export 'src/persistence_local/persistence_local_adapter_base.dart';
+export 'src/persistence_local/persistence_local_adapter_factory_base.dart';
+export 'src/persistence_local/persistence_local_provider_base.dart';
