@@ -2,7 +2,7 @@ import 'package:core_persistence_base/core_persistence_base.dart';
 
 import '../core_persistence_local.dart';
 
-abstract interface class CorePersistenceProviderLocal
+abstract interface class CorePersistenceLocalProvider
     extends PersistenceProviderBase {
-  CorePersistenceAdapterLocalFactory getAdapterFactory(String persistenceId);
+  CorePersistenceLocalAdapterFactory getAdapterFactory(String persistenceId);
 }
