@@ -1,12 +1,16 @@
 import 'dart:async';
 
 import 'package:core_object/core_object.dart';
+import 'package:core_persistence_local/core_persistence_local.dart';
 
-abstract class CorePersistenceAdapterRemote {
-  // ifEmpty == null - expect that remote is initialized
+abstract class PersistenceLocalAdapterBase
+    implements CorePersistenceLocalAdapter {
+
+  @override
   Future<({String ref, int sequenceNumber})?> initialize({
     required ({
       String ref,
+      int sequenceNumber,
       int createdAt,
       StateViewObject? stateViewData,
     })
@@ -14,12 +18,18 @@ abstract class CorePersistenceAdapterRemote {
         ifEmpty,
   });
 
+  @override
   Future<void> add({
     Map<String, ({String ref, Iterable<String> parent, int createdAt})>? entry,
     Map<String, JsonMap>? event,
     Map<String, StateViewObject>? stateView,
   });
 //
+
+}
+
+
+
 // Future<void> append({
 //   required JsonMap entry,
 //   required JsonMap? event,
@@ -38,4 +48,3 @@ abstract class CorePersistenceAdapterRemote {
 // Stream<Map<Ref, JsonMap>> get eventAll;
 //
 // Stream<Map<Ref, StateViewData>> get stateViewAll;
-}
