@@ -2,9 +2,9 @@ import 'package:meta/meta.dart';
 
 import '../../core_persistence_local.dart';
 
-abstract class PersistenceAdapterLocalFactoryBase
-    implements CorePersistenceAdapterLocalFactory {
-  PersistenceAdapterLocalFactoryBase({
+abstract class PersistenceLocalAdapterFactoryBase
+    implements CorePersistenceLocalAdapterFactory {
+  PersistenceLocalAdapterFactoryBase({
     required String persistenceId,
   }) : _persistenceId = persistenceId;
 
