@@ -3,7 +3,7 @@ import 'package:core_persistence_base/core_persistence_base.dart';
 import 'package:core_persistence_local/core_persistence_local.dart';
 import 'package:sembast/sembast.dart';
 
-class PersistenceAdapterLocalSembast implements PersistenceAdapterLocalBase {
+class PersistenceAdapterLocalSembast implements PersistenceLocalAdapterBase {
   PersistenceAdapterLocalSembast({
     required String persistenceId,
     required Database database,
