@@ -3,7 +3,7 @@ import 'package:core_persistence_local_sembast/core_persistence_local_sembast.da
 import 'package:sembast/sembast.dart';
 import 'package:sembast/sembast_memory.dart';
 
-class PersistenceProviderLocalSembast implements PersistenceProviderLocalBase {
+class PersistenceProviderLocalSembast implements PersistenceLocalProviderBase {
   PersistenceProviderLocalSembast({
     required this.databaseFactory,
   });
@@ -14,7 +14,7 @@ class PersistenceProviderLocalSembast implements PersistenceProviderLocalBase {
   final DatabaseFactory databaseFactory;
 
   @override
-  CorePersistenceAdapterLocalFactory getAdapterFactory(String persistenceId) =>
+  CorePersistenceLocalAdapterFactory getAdapterFactory(String persistenceId) =>
       PersistenceAdapterLocalFactorySembast(
         persistenceId: persistenceId,
         databaseFactory: databaseFactory,
