@@ -1,14 +1,14 @@
-import 'package:core_persistence_remote/src/persistence_adapter_remote.dart';
+import 'package:core_persistence_remote/src/core_persistence_remote_adapter.dart';
 import 'package:meta/meta.dart';
 
 
-abstract class CorePersistenceAdapterRemoteFactory {
-  Future<CorePersistenceAdapterRemote> getAdapter(String path);
+abstract class CorePersistenceRemoteAdapterFactory {
+  Future<CorePersistenceRemoteAdapter> getAdapter(String path);
 }
 
 // - Base
 abstract class PersistenceAdapterRemoteFactoryBase
-    implements CorePersistenceAdapterRemoteFactory {
+    implements CorePersistenceRemoteAdapterFactory {
   PersistenceAdapterRemoteFactoryBase({
     required String persistenceId,
   }) : _persistenceId = persistenceId;
