@@ -1,29 +1,27 @@
 import 'dart:async';
 
 import 'package:core_object/core_object.dart';
-import 'package:core_persistence_local/core_persistence_local.dart';
+import 'package:core_persistence_base/src/object_instance_head.dart';
+import 'package:core_persistence_remote/core_persistence_remote.dart';
 
-abstract class PersistenceAdapterLocalBase
-    implements CorePersistenceAdapterLocal {
-  @override
-  Future<({String ref, int sequenceNumber})?> initialize({
-    required ({
-      String ref,
-      int sequenceNumber,
-      int createdAt,
-      StateViewObject? stateViewData,
-    })
-            Function()?
-        ifEmpty,
-  });
-
+abstract class PersistenceRemoteAdapterBase
+    implements CorePersistenceRemoteAdapter {
   @override
   Future<void> add({
     Map<String, ({String ref, Iterable<String> parent, int createdAt})>? entry,
     Map<String, JsonMap>? event,
     Map<String, StateViewObject>? stateView,
   });
+
+  @override
+  Future<ObjectInstanceHead> initialize(
+      {required IfEmptyCallback? ifEmpty,}) {
+    throw UnimplementedError();
+  }
 //
+}
+
+
 // Future<void> append({
 //   required JsonMap entry,
 //   required JsonMap? event,
@@ -42,4 +40,3 @@ abstract class PersistenceAdapterLocalBase
 // Stream<Map<Ref, JsonMap>> get eventAll;
 //
 // Stream<Map<Ref, StateViewData>> get stateViewAll;
-}
