@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:core_object/core_object.dart';
-import 'package:core_object_store_remote/src/object_store_remote/object_store_remote_state.dart';
 import 'package:core_persistence_remote/core_persistence_remote.dart';
 
 import '../../core_object_store_remote.dart';
@@ -11,14 +10,39 @@ class ObjectStoreRemote extends Cubit<ObjectStoreRemoteState>
     implements CoreObjectStoreRemote {
   ObjectStoreRemote(
     super.initialState, {
-    required CorePersistenceAdapterRemote adapter,
+    required CorePersistenceRemoteAdapter adapter,
   }) : _adapter = adapter {
-    _effect.stream.listen((event) {
-      _update.add(ObjectUpdateRemote.initial(ref: '1', sequenceNumber: 1));
+    _effect.stream.listen((event) async {
+      await event.map(
+        initialize: (initialize) async {
+          await state.map(
+            initial: (initial) async {
+              final result = await _adapter.initialize(
+                ifEmpty: initialize.ifEmpty,
+              );
+              _update.add(
+                ObjectUpdateRemote.initial(
+                  ref: result.ref,
+                  sequenceNumber: result.sequenceNumber,
+                ),
+              );
+              emit(ObjectStoreRemoteState.ready());
+            },
+            ready: (ready) {
+              addError(UnsupportedError('instance already initialized'));
+            },
+          );
+        },
+        append: (append) {},
+        forward: (forward) {},
+        publish: (publish) {},
+        none: (none) {},
+      );
+      // _update.add(ObjectUpdateRemote.initial(ref: '1', sequenceNumber: 1));
     });
   }
 
-  final CorePersistenceAdapterRemote _adapter;
+  final CorePersistenceRemoteAdapter _adapter;
 
   @override
   StreamSink<ObjectEffectRemote> get effect => _effect.sink;
