@@ -9,14 +9,12 @@
 // coverage:ignore-file
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:core_object_store_remote/core_object_store_remote.dart' as _i4;
 import 'package:core_object_store_remote_test/core_object_store_remote_test.dart'
     as _i3;
-import 'package:core_persistence_remote/core_persistence_remote.dart' as _i5;
 import 'package:get_it/get_it.dart' as _i1;
 import 'package:injectable/injectable.dart' as _i2;
 
-import 'configure.dart' as _i6;
+import 'configure.dart' as _i4;
 
 extension GetItInjectableX on _i1.GetIt {
 // initializes the registration of main-scope dependencies inside of GetIt
@@ -31,10 +29,9 @@ extension GetItInjectableX on _i1.GetIt {
     );
     await _i3.CoreObjectStoreRemoteTestPackageModule().init(gh);
     final module = _$Module();
-    gh.factory<_i4.ObjectStoreRemote>(
-        () => module.getObjectStoreRemote(gh<_i5.CorePersistenceAdapterRemote>()));
+    gh.factory<String>(() => module.hello);
     return this;
   }
 }
 
-class _$Module extends _i6.Module {}
+class _$Module extends _i4.Module {}
