@@ -1,6 +1,4 @@
-import 'package:core_object_store_remote/core_object_store_remote.dart';
 import 'package:core_object_store_remote_test/core_object_store_remote_test.dart';
-import 'package:core_persistence_remote/core_persistence_remote.dart';
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 
@@ -19,9 +17,10 @@ Future<void> configureDependencies() async => getIt.init();
 
 @module
 abstract class Module {
-  ObjectStoreRemote getObjectStoreRemote(CorePersistenceAdapterRemote adapter) =>
-      ObjectStoreRemote(
-        ObjectStoreRemoteState.initial(),
-        adapter: adapter,
-      );
+  String get hello => 'hi';
+  // ObjectStoreRemote getObjectStoreRemote(CorePersistenceAdapterRemote adapter) =>
+  //     ObjectStoreRemote(
+  //       ObjectStoreRemoteState.initial(),
+  //       adapter: adapter,
+  //     );
 }
